+++ changelog.md
@@ -1,1 +1,2 @@
-- 【2024/11/25】Finish chapter_preliminaries\my-preliminaries.ipynb
+- 【2024/11/25】Finish chapter_preliminaries\my-preliminaries.ipynb 
+- 【2024/11/26】Finish chapter_linear-networks\my-linear-regression.ipynb & my-linear-regression-scratch-2.ipynb &    my-linear-regression-concise-4.py